Updated daily-build scripts.

+++ install-scripts/mac/bundle-daily-build.sh
@@ -19,7 +19,11 @@ fi
 
 # Defaults are current latest revisions in stable branch and trunk
 STABLE_REVISION=${1:-`svn info --non-interactive http://www.ailab.si/svn/orange/branches/ver1.0/ | grep 'Last Changed Rev:' | cut -d ' ' -f 4`}
+# svn info does not return proper exit status on an error so we check it this way
+[ $STABLE_REVISION ] && exit 3
 DAILY_REVISION=${2:-`svn info --non-interactive http://www.ailab.si/svn/orange/trunk/ | grep 'Last Changed Rev:' | cut -d ' ' -f 4`}
+# svn info does not return proper exit status on an error so we check it this way
+[ $DAILY_REVISION ] && exit 4
 
 echo "Preparing temporary directory."
 rm -rf /private/tmp/bundle/
